Do not filter out 'this' variables in static methods
Kotlin generates some lambdas as static methods with 'this' argument. Without this fix 'this' argument is not visible. It might be possible to extract 'this' by argument value. But it doesn't work in the case when some lambda parameter is anonymous: JDWP protocol error will be thrown in the getArgumentValues call (see 8323747:).